Resize web tables if number of "objects" changes
When number of elements in table (acquired via API) changes, then
clear and redraw table on the page.

This fixes the bug when number of fans or pools changed and the page
froze. This was because the code was trying to access non-existent
page elements and got error when doing so.