web gui: fix problems with "empty-values"
The problem in a nutshell is: luci/cbi doesn't seem to support
non-optional dependent values.

That means that if value is to be dependent on some other values (for
example "overclocking enabled" with dependent "overclock" value), the
overclock value cannot be non-optional and non-removable, because
otherwise error shows up even when dependencies are not met.

This patch just checks all numeric values are indeed numeric.